some code modifications

+++ synchronized-commons/src/main/java/fr/insalyon/creatis/vip/synchronizedcommons/SyncedDevice.java
@@ -1,44 +1,43 @@
 /*
-Copyright 2015
+ Copyright 2015
 
-CREATIS
-CNRS UMR 5220 -- INSERM U1044 -- Université Lyon 1 -- INSA Lyon
+ CREATIS
+ CNRS UMR 5220 -- INSERM U1044 -- Université Lyon 1 -- INSA Lyon
 
-Authors
+ Authors
 
-Nouha Boujelben (nouha.boujelben@creatis.insa-lyon.fr)
-Tristan Glatard (tristan.glatard@creatis.insa-lyon.fr)
+ Nouha Boujelben (nouha.boujelben@creatis.insa-lyon.fr)
+ Tristan Glatard (tristan.glatard@creatis.insa-lyon.fr)
 
-This software is a daemon for file synchronization between SFTP
-servers and the LCG File Catalog (LFC).
+ This software is a daemon for file synchronization between SFTP
+ servers and the LCG File Catalog (LFC).
 
-This software is governed by the CeCILL-B license under French law and
-abiding by the rules of distribution of free software.  You can use,
-modify and/ or redistribute the software under the terms of the
-CeCILL-B license as circulated by CEA, CNRS and INRIA at the following
-URL "http://www.cecill.info".
+ This software is governed by the CeCILL-B license under French law and
+ abiding by the rules of distribution of free software.  You can use,
+ modify and/ or redistribute the software under the terms of the
+ CeCILL-B license as circulated by CEA, CNRS and INRIA at the following
+ URL "http://www.cecill.info".
 
-As a counterpart to the access to the source code and rights to copy,
-modify and redistribute granted by the license, users are provided
-only with a limited warranty and the software's author, the holder of
-the economic rights, and the successive licensors have only limited
-liability.
+ As a counterpart to the access to the source code and rights to copy,
+ modify and redistribute granted by the license, users are provided
+ only with a limited warranty and the software's author, the holder of
+ the economic rights, and the successive licensors have only limited
+ liability.
 
-In this respect, the user's attention is drawn to the risks associated
-with loading, using, modifying and/or developing or reproducing the
-software by the user in light of its specific status of free software,
-that may mean that it is complicated to manipulate, and that also
-therefore means that it is reserved for developers and experienced
-professionals having in-depth computer knowledge. Users are therefore
-encouraged to load and test the software's suitability as regards
-their requirements in conditions enabling the security of their
-systems and/or data to be ensured and, more generally, to use and
-operate it in the same conditions as regards security.
+ In this respect, the user's attention is drawn to the risks associated
+ with loading, using, modifying and/or developing or reproducing the
+ software by the user in light of its specific status of free software,
+ that may mean that it is complicated to manipulate, and that also
+ therefore means that it is reserved for developers and experienced
+ professionals having in-depth computer knowledge. Users are therefore
+ encouraged to load and test the software's suitability as regards
+ their requirements in conditions enabling the security of their
+ systems and/or data to be ensured and, more generally, to use and
+ operate it in the same conditions as regards security.
 
-The fact that you are presently reading this means that you have had
-knowledge of the CeCILL-B license and that you accept its terms.
-*/
-
+ The fact that you are presently reading this means that you have had
+ knowledge of the CeCILL-B license and that you accept its terms.
+ */
 package fr.insalyon.creatis.vip.synchronizedcommons;
 
 import fr.insalyon.creatis.vip.synchronizedcommons.business.SyncException;
@@ -186,8 +185,8 @@ public interface SyncedDevice {
      */
     public double getNbSecondFromConfigFile() throws SyncException;
 
-    public void updateLFCMonitoringParams(Synchronization ua, int numberOfFilesTransferredToLFC, double sizeOfFilesTransferredToLFC, int numberOfFilesDeletedInLFC, double sizeOfFilesDeletedInLFC) throws SyncException;
+    public void updateLFCMonitoringParams(Synchronization ua, int numberOfFilesTransferredToLFC, long sizeOfFilesTransferredToLFC, int numberOfFilesDeletedInLFC, long sizeOfFilesDeletedInLFC) throws SyncException;
 
-    public void updateDeviceMonitoringParams(Synchronization ua, int numberOfFilesTransferredToDevice, double sizeOfFilesTransferredToDevice, int numberOfFilesDeletedInDevice, double sizeOfFilesDeletedInDevice) throws SyncException;
+    public void updateDeviceMonitoringParams(Synchronization ua, int numberOfFilesTransferredToDevice, long sizeOfFilesTransferredToDevice, int numberOfFilesDeletedInDevice, long sizeOfFilesDeletedInDevice) throws SyncException;
 
 }
